+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 
+- Allow for offsets in OMETiffLoader for `bioformats6` pyramids.
 - Update geotiff off from Ilan's for to the official release.
 
 ## 0.2.0